向量資料庫_rag/vector_db_project/requirements.txt
streamlit>=1.24.0
python-dotenv>=1.0.0
google-api-python-client>=2.0.0
llama-index>=0.10.0
google-generative-ai>=0.3.0
chromadb>=0.4.0
doc